requirements.txt
django
djangorestframework
requests
pymysql
cryptography
python-dotenv